+++ src/Logging.ts
@@ -12,8 +12,22 @@ export interface ErrorLoggerOptions {
   providedInfo: any;
 }
 
-export const defaultErrorLogger: ErrorLogger = ({source, message, error, info}) => {
-  console.error(source || `SendableError`, message, info, error.stack);
+export const defaultErrorLogger: ErrorLogger = (options) => {
+  const {source, message, error, info} = options;
+
+  const stackLines = error.stack?.split("\n").slice(1).join("\n");
+  console.error(`${source || `SendableError`} - ${message} ${JSON.stringify(info || {})}${stackLines ? `\n${stackLines}` : ""}`);
+
+  const cause = error.cause;
+  if (cause && cause instanceof Error) {
+    defaultErrorLogger({
+      ...options,
+      source,
+      message  : `Caused by: ${cause.message}`,
+      error    : cause as SendableError,
+      errorInfo: {},
+    });
+  }
 };
 
 let currentLogger = defaultErrorLogger;